After hours of debugging, extracted the email validation feature into its own file
