+++ app/static/js/requestOtherEngagement.js
@@ -1,6 +1,9 @@
+import { validateEmail } from "./emailValidation.mjs";
+
 $(document).ready(function(e) {
 
-    $("input.phoneInput").inputmask('(999)-999-9999');
+    $("input.phone-input").inputmask('(999)-999-9999');
+    $("input.email-input").on('input', validateEmail)
 
 
 });
