+++ app/templates/main/emergencyContactInfo.html
@@ -4,7 +4,7 @@
 {% block scripts %}
     {{super()}}
     <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery.inputmask/3.3.4/jquery.inputmask.bundle.min.js"></script>
-    <script type="text/javascript" src="{{url_for('static', filename='js/emergencyContactInfo.js') }}?u={{lastStaticUpdate}}"></script>
+    <script type="module" src="{{url_for('static', filename='js/emergencyContactInfo.js') }}?u={{lastStaticUpdate}}"></script>
 {% endblock %}
 
 {% block app_content %}
@@ -31,7 +31,7 @@
             </div>
             <div class='form-group mb-4'>
                 <label class='form-label' for='email' ><strong> Email Address </strong></label>
-                <input {{'disabled' if readOnly else ''}} class='form-control' data-inputmask-alias="email" type='email' id='emergencyContactEmail' value="{{contactInfo.emailAddress if contactInfo else ''}}" pattern="[a-z0-9._%+-]+@[a-z0-9.-]+\.[a-z]{2,4}$" name='emailAddress' />
+                <input {{'disabled' if readOnly else ''}} class='form-control' type='text' id='emergencyContactEmail' value="{{contactInfo.emailAddress if contactInfo else ''}}" name='emailAddress'/>
             </div>
 
         </div>
