+++ app/templates/minor/requestOtherEngagement.html
@@ -4,6 +4,8 @@
 {% block scripts %}
     {{super()}}
     <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
+    <!-- Beans: Is this a very good idea? To pull a plugin from a cloudflare server like this? We do this in emergencyContaceInfo.html as well -->
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery.inputmask/5.0.7/jquery.inputmask.min.js"></script>  
     <script type="module" src="/static/js/displayFilesMacro.js"></script>
     <script type="module" src="/static/js/requestOtherEngagement.js"></script>
     <script src="//cdnjs.cloudflare.com/ajax/libs/timepicker/1.3.5/jquery.timepicker.min.js"></script>
@@ -57,10 +59,9 @@
       <br>
 
       <div class="row">
-        <!-- Beans: We can do more to validate the form of this input. I believe it's a javascript function with an input mask-->
         <div class="form-group col-6">
           <label class="form-label" for="companyPhone"> <strong>Phone Number</strong> </label>
-          <input class="form-control phoneInput" id="companyPhone" placeholder="Company or organization's phone number" name="companyPhone"/>
+          <input class="form-control phone-input" id="companyPhone" placeholder="Company or organization's phone number" name="companyPhone"/>
         </div>
 
         <div class="form-group col-6">
@@ -76,15 +77,14 @@
       <div>
 
       <div class="row">
-        <!-- Beans: We can do more to validate the form of this input. I believe it's a javascript function with an input mask-->
         <div class="form-group col-6">
           <label class="form-label" for="supervisorPhone"> <strong>Phone Number</strong> </label>
-          <input class="form-control phoneInput" id="supervisorPhone" placeholder="Supervisor's phone number" name="supervisorPhone"/>
+          <input class="form-control phone-input" id="supervisorPhone" placeholder="Supervisor's phone number" name="supervisorPhone"/>
         </div>
 
         <div class="form-group col-6">
           <label class="form-label" for="supervisorEmail"> <strong>Email</strong> </label>
-          <input class="form-control" id="supervisorEmail" placeholder="Supervisor's email" name="supervisorEmail"/>
+          <input class="form-control email-input" id="supervisorEmail" placeholder="Supervisor's email" name="supervisorEmail"/>
         </div>
       </div>
